Update The Circle Trail Blazer
(Makiese Light Sita - Tuesday, September 21 2021)
This is a Circle Trail Blazer! It allows you to control a red dot that shrinks and grows with the hold of a mouse button.  If the dot shrinks, you can press any key of your choosing to have it grow. If the dot grows, you can also press any key of your choice in order to shrink it. The challenge is to keep the red dot at a relative size consistently as long as you can. HAVE FUN!!!