Add files via upload
This is a program created by me, which serves as a two minute countdown. Each minute, the color of the clock, as well as the color of the messages that pop up telling you how many minutes are left, change colors in perfect harmony. To start the clock, you need to hold a key. To cancel it completely, you need to release the key that you have pressed. The title cards have been done by me, Makiese Light Sita, while the background has been imported from the internet. HAVE FUN!!!