Add files via upload
(Tuesday, December 7 2021)
This is a MIDI color changer program. With the press of 4 different buttons, you have the power to edit the pitch of the MIDI piano sound as well as the colors that each pitch triggers onto the circle that you see in the screen. The 'z' key turns the circle yellow, the 'x' key turns the circle purple, the 'c' key turns the circle light blue, and the 'v' key turns the circle green. HAVE FUN!!!